unities/requirements.txt
psycopg2-binary==2.9.3
sanic==22.6.2
sanic-openapi==21.12.0
pydantic~=1.10.2
requests~=2.28.1